requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.2.9
gunicorn==20.0.4
djangorestframework==3.11.0
django-cors-headers==3.7.0
djongo==1.3.6
dnspython==2.1.0
pymongo==3.12.1
django-environ==0.8.1
djangorestframework-simplejwt==5.0.0
pygments==2.10.0
mango-jwt==1.3.2
PyJWT